+++ m365-apps/repair-and-trigger-m365-apps-updates/Detect.ps1
@@ -3,6 +3,7 @@
 # Script Name:         Detect.ps1
 # Description:         Detect if Office is installed, if "Click to Run Service" is running and trigger update
 # Changelog:           2023-11-14: Initial version
+#                      2025-01-17: Added reset of UpdateDetectionLastRunTime
 #
 #=============================================================================================================================
 
@@ -10,8 +11,26 @@
 $curSvcStat, $svcCTRSvc, $errMsg = "", "", ""
 $OfficeC2RClientPath = Join-Path $env:ProgramFiles "\Common Files\microsoft shared\ClickToRun\"
 $OfficeC2RClientProcess = Join-Path $OfficeC2RClientPath "OfficeC2RClient.exe"
-$OfficeC2RClientArgs = "/frequentupdate SCHEDULEDTASK displaylevel=False"
+$OfficeC2RClientArgs = "/update user forceappshutdown=false displaylevel=false"
 
+$pathUpdateDetectionLastRunTime = "HKLM:\SOFTWARE\Microsoft\Office\ClickToRun\Updates\"
+$nameUpdateDetectionLastRunTime = "UpdateDetectionLastRunTime"
+
+# functions
+Function Test-RegistryValue {
+    param (
+        [parameter(Mandatory=$true)][ValidateNotNullOrEmpty()]$Path,
+        [parameter(Mandatory=$true)][ValidateNotNullOrEmpty()]$Value
+    )
+    
+    try {
+      Get-ItemProperty -Path $Path -ErrorAction Stop | Select-Object -ExpandProperty $Value -ErrorAction Stop | Out-Null
+      return $true
+    } catch {
+      return $false
+    }
+  
+}
 
 if (-not (Test-Path -Path 'HKLM:\Software\Microsoft\Office\16.0')) {
     Write-Host "Office 16.0 (or greater) not present on this machine"
@@ -29,10 +48,14 @@ try {
 
 if ($curSvcStat -eq "Running") {
     Write-Host $curSvcStat
-    if (Test-Path $OfficeC2RClientProcess ) {
+    if (Test-Path $OfficeC2RClientProcess) {
+        # reset UpdateDetectionLastRunTime to let service search for updates directly
+        if(Test-RegistryValue $pathUpdateDetectionLastRunTime $nameUpdateDetectionLastRunTime) {
+            Remove-ItemProperty -Path $pathUpdateDetectionLastRunTime -Name $nameUpdateDetectionLastRunTime -Force | Out-Null
+        }
+        # start OfficeC2RClientProcess
         Set-Location -Path $OfficeC2RClientPath
         Start-Process $OfficeC2RClientProcess $OfficeC2RClientArgs
-        Write-Host "Triggered Office C2R Client Process"
         exit 0
     }
     else {